Move NASA testing functions to Helper class
This way, we can easily reuse these functions.
