+++ test/Makefile.am
@@ -128,6 +128,7 @@ pkginclude_HEADERS += standard_unit/reaction_rate_test_base.h
 pkginclude_HEADERS += standard_unit/reaction_rate_vector_test_base.h
 pkginclude_HEADERS += standard_unit/arrhenius_rate_test_helper.h
 pkginclude_HEADERS += standard_unit/arrhenius_rate_vector_test_base.h
+pkginclude_HEADERS += standard_unit/nasa_test_helper.h
 pkginclude_HEADERS += standard_unit/nasa7_thermo_test_base.h
 pkginclude_HEADERS += standard_unit/nasa9_thermo_test_base.h
 pkginclude_HEADERS += standard_unit/thermo_vector_test_base.h
