+++ test/standard_unit/nasa7_thermo_test_base.h
@@ -30,10 +30,13 @@
 // Antioch
 #include "antioch/cmath_shims.h"
 
+// Antioch Testing
+#include "nasa_test_helper.h"
+
 namespace AntiochTesting
 {
   template<typename Scalar>
-  class NASA7ThermoTestBase
+  class NASA7ThermoTestBase : public NASA7ThermoTestHelper<Scalar>
   {
   public:
 
@@ -46,21 +49,6 @@ namespace AntiochTesting
       this->init_N2_coeffs_300_1000();
       this->init_N2_coeffs_1000_5000();
       this->init_all_N2_coeffs();
-    }
-
-    Scalar cp_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2, Scalar a3, Scalar a4 )
-    {
-      return a0 + a1*T + a2*T*T + a3*T*T*T + a4*(T*T*T*T);
-    }
-
-    Scalar h_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2, Scalar a3, Scalar a4, Scalar a5 )
-    {
-      return a0 + a1/2.0L*T + a2/3.0L*T*T + a3/4.0L*T*T*T + a4/5.0L*(T*T*T*T) + a5/T;
-    }
-
-    Scalar s_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2, Scalar a3, Scalar a4, Scalar a6 )
-    {
-      return a0*std::log(T) + a1*T + a2/2.0L*T*T + a3/3.0L*T*T*T + a4/4.0L*(T*T*T*T) + a6;
     }
 
     void init_H2_coeffs_200_1000()
