+++ test/standard_unit/nasa9_thermo_test_base.h
@@ -30,10 +30,13 @@
 // Antioch
 #include "antioch/cmath_shims.h"
 
+// Antioch Testing
+#include "nasa_test_helper.h"
+
 namespace AntiochTesting
 {
   template<typename Scalar>
-  class NASA9ThermoTestBase
+  class NASA9ThermoTestBase : public NASA9ThermoTestHelper<Scalar>
   {
   public:
 
@@ -53,27 +56,7 @@ namespace AntiochTesting
       this->init_all_NO2_coeffs();
     }
 
-    Scalar cp_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2,
-                     Scalar a3, Scalar a4, Scalar a5, Scalar a6 )
-    {
-      return a0/(T*T) + a1/T + a2 + a3*T + a4*(T*T) + a5*(T*T*T) + a6*(T*T*T*T);
-    }
 
-    Scalar h_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2,
-                    Scalar a3, Scalar a4, Scalar a5, Scalar a6,
-                    Scalar a7 )
-    {
-      return -a0/(T*T) + a1*std::log(T)/T + a2 + a3*T/2.0L + a4*(T*T)/3.0L
-        + a5*(T*T*T)/4.0L + a6*(T*T*T*T)/5.0L + a7/T;
-    }
-
-    Scalar s_exact( Scalar T, Scalar a0, Scalar a1, Scalar a2,
-                    Scalar a3, Scalar a4, Scalar a5, Scalar a6,
-                    Scalar a8 )
-    {
-      return -a0/(2.L*T*T) - a1/T + a2*std::log(T) + a3*T + a4*(T*T)/2.0L
-        + a5*(T*T*T)/3.0L + a6*(T*T*T*T)/4.0L + a8;
-    }
 
     void init_N2_coeffs_0_200()
     {
